requirements.txt
streamlit
flask
google-auth
google-auth-oauthlib
google-api-python-client
sentence-transformers
spacy
transformers
